SWDEV-296922 : Incorrect rounding due to integer division in rocprofiler metrics
Changed derived metrics to double from int64.
Fixed standalone test due to int64 to float change
Fixed intercept test due to int64 to float change.



[ROCm/rocprofiler commit: f9017cbdc57166c417474b6794bfc08b703e356a]

+++ src/xml/expr.h
@@ -29,6 +29,7 @@ THE SOFTWARE.
 #include <iostream>
 #include <sstream>
 #include <string.h>
+#include <float.h>
 
 namespace xml {
 class exception_t : public std::exception {
@@ -45,8 +46,8 @@ class div_zero_exception_t : public exception_t {
   explicit div_zero_exception_t(const std::string& msg) : exception_t("Divide by zero exception " + msg) {}
 };
 
-typedef uint64_t args_t;
-static const args_t ARGS_MAX = UINT64_MAX;
+typedef double args_t;
+static const args_t ARGS_MAX = DBL_MAX;
 typedef std::map<std::string, args_t> args_map_t;
 class Expr;
